data_management: refactoring 2
